requirements.txt
Django==4.0
typing_extensions==4.0.1
whitenoise==5.3.0
djangorestframework==3.13.1